requirements.txt
alembic==1.14.1
anytree==2.12.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==41.0.3
dnspython==2.3.0
Flask==2.3.3
Flask-JWT-Extended==4.5.2
flask_migrate==4.0.5
Flask-SQLAlchemy==3.1.1
future-fstrings==1.2.0
greenlet==3.1.1
gunicorn==20.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==3.0.2
merklelib==1.0
pip==25.0.1
pycparser==2.22
PyJWT==2.10.1
pyperclip==1.8.2
python-dotenv==1.0.0
setuptools==75.8.0
six==1.17.0
SQLAlchemy==2.0.21
sqlparse==0.4.4
typing_extensions==4.12.2
urllib3==2.0.4
Werkzeug==2.3.7
PyNaCl==1.5.0
base58==2.1.1
requests==2.32.3
